refactor : js in ts
