+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+      "allowJs": true,
+      "target": "ES5",
+      "outDir": "./dist",
+      "moduleResolution": "Node",
+      "lib": ["ES2015", "DOM", "DOM.Iterable"]
+    },
+    "include": ["./src/**/*"],
+    "exclude": ["node_modules", "dist"]
+  }
